@@ -1,9 +1,9 @@
 package main
 
 import (
+	"bytes"
 	"crypto/rand"
 	"encoding/base64"
-	"errors"
 	"fmt"
 	"strings"
 	"sync"
@@ -14,22 +14,16 @@ import (
 )
 
 // shortTokenD - debugging ShortToken interface realization
-type shortTokenD struct {
-	length int
-}
+type shortTokenD struct{ length int }
 
 // mockShortToken returns the shortToken interface that always returns the same token
-func mockShortToken(length int) ShortToken {
-	return &shortTokenD{length}
-}
+func mockShortToken(length int) ShortToken { return &shortTokenD{length} }
 
-func (s shortTokenD) Get() string {
-	return strings.Repeat("_", s.length)
-}
+func (s shortTokenD) Get() string { return strings.Repeat("_", s.length) }
 
-func (s shortTokenD) Check(_ string) error {
-	return nil
-}
+func (s shortTokenD) CheckAlphabet(_ string) error { return nil }
+
+func (s shortTokenD) CheckLength(_ string) error { return nil }
 
 // try to create new token from debugging source
 func TestMockShortToken(t *testing.T) {
@@ -37,32 +31,40 @@ func TestMockShortToken(t *testing.T) {
 	require.Equal(t, strings.Repeat("_", 6), st.Get())
 }
 
+const urlEncodingString = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"
+
+func TestCheckAlphabet(t *testing.T) {
+	st := NewShortToken(8)
+	require.NoError(t, st.CheckAlphabet(urlEncodingString))
+	for _, ch := range "~`!№@#$%^&*/+=(){}[]<>?,.\"':;|\\" {
+		assert.Error(t, st.CheckAlphabet(string(ch)))
+	}
+}
+
 // try to make two tokens from random source and compare them
 func TestNewShortTokenReal6(t *testing.T) {
 	st := NewShortToken(6)
-
 	tc := st.Get()
-
 	tc1 := st.Get()
-
-	require.Equal(t, 6, len(tc))
-	require.Equal(t, 6, len(tc1))
-	require.NotEqual(t, tc, tc1)
 	t.Log(tc, tc1)
+	require.NoError(t, st.CheckLength(tc))
+	require.NoError(t, st.CheckLength(tc1))
+	require.NoError(t, st.CheckAlphabet(tc), tc)
+	require.NoError(t, st.CheckAlphabet(tc1), tc1)
+	require.NotEqual(t, tc, tc1)
 }
 
 // try to make two very short tokens from random source and compare them
 func TestNewShortTokenReal2(t *testing.T) {
 	st := NewShortToken(2)
-
 	tc := st.Get()
-
 	tc1 := st.Get()
-
-	require.Equal(t, 2, len(tc))
-	require.Equal(t, 2, len(tc1))
-	require.NotEqual(t, tc, tc1)
 	t.Log(tc, tc1)
+	require.NoError(t, st.CheckLength(tc))
+	require.NoError(t, st.CheckLength(tc1))
+	require.NoError(t, st.CheckAlphabet(tc), tc)
+	require.NoError(t, st.CheckAlphabet(tc1), tc1)
+	require.NotEqual(t, tc, tc1)
 }
 
 func TestAsyncGet(t *testing.T) {
@@ -88,32 +90,23 @@ func TestAsyncGet(t *testing.T) {
 // test Check with correct token
 func Test00ST10CheckOk(t *testing.T) {
 	st := NewShortToken(2)
-
 	sToken := st.Get()
-
-	require.NoError(t, st.Check(sToken))
+	require.NoError(t, st.CheckLength(sToken))
+	require.NoError(t, st.CheckAlphabet(sToken))
 }
 
 // test Check with wrong token length
 func Test00ST15CheckNoOk(t *testing.T) {
 	st := NewShortToken(2)
-
 	sToken := st.Get()
-
-	require.Error(t, st.Check(sToken+"wrong"))
-}
-
-// test Check with wrong token alphabet
-func Test00ST15CheckNoOk2(t *testing.T) {
-	st := NewShortToken(2)
-
-	require.Error(t, st.Check("#$")) // check nonBase64 symbols
+	require.Error(t, st.CheckLength(sToken+"wrong"))
+	require.Error(t, st.CheckAlphabet(sToken+"!@#"))
 }
 
 // Benchmark for the 2 symbols token
 func Benchmark00ST00Create2(b *testing.B) {
 	st := NewShortToken(2)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
@@ -121,7 +114,7 @@ func Benchmark00ST00Create2(b *testing.B) {
 // Benchmark for the 6 symbols token
 func Benchmark00ST00Create6(b *testing.B) {
 	st := NewShortToken(6)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
@@ -129,7 +122,7 @@ func Benchmark00ST00Create6(b *testing.B) {
 // Benchmark for the 8 symbols token
 func Benchmark00ST00Create8(b *testing.B) {
 	st := NewShortToken(8)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
@@ -138,7 +131,7 @@ func Benchmark00ST00Create8(b *testing.B) {
 // unfortunately it is slower and requires more memory than original simple version.
 func NewBShortToken(length int) ShortToken {
 	pool := &sync.Pool{}
-	pool.New = func() interface{} {
+	pool.New = func() any {
 		return make([]byte, length*6/8+1)
 	}
 	return &shortBToken{
@@ -152,7 +145,7 @@ type shortBToken struct {
 	bufPool *sync.Pool // buffer pool for random bytes
 }
 
-// Get creates the token from random or debugging source
+// Get creates the token from random source
 func (s *shortBToken) Get() string {
 	// get secure random bytes
 	buf := s.bufPool.Get().([]byte)
@@ -166,14 +159,14 @@ func (s *shortBToken) Get() string {
 	return base64.URLEncoding.EncodeToString(buf)[:s.length]
 }
 
-func (s *shortBToken) Check(sToken string) error {
-	return nil
-}
+func (s *shortBToken) CheckLength(sToken string) error { return nil }
+
+func (s *shortBToken) CheckAlphabet(sToken string) error { return nil }
 
 // Benchmark for the 2 symbols token
 func Benchmark00ST00Create2B(b *testing.B) {
 	st := NewBShortToken(2)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
@@ -181,7 +174,7 @@ func Benchmark00ST00Create2B(b *testing.B) {
 // Benchmark for the 6 symbols token
 func Benchmark00ST00Create6B(b *testing.B) {
 	st := NewBShortToken(6)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
@@ -189,51 +182,81 @@ func Benchmark00ST00Create6B(b *testing.B) {
 // Benchmark for the 8 symbols token
 func Benchmark00ST00Create8B(b *testing.B) {
 	st := NewBShortToken(8)
-	for range b.N {
+	for b.Loop() {
 		_ = st.Get()
 	}
 }
 
-var (
-	tokens = []string{"ABCDEFGH", "ABCDEFG&", "ABCDEF&G", "ABCDE&FG", "ABCD&EFG", "&ABCDEFG", "ABCDEFGH"}
-)
+var tokens = []string{"ABCDEFGH", "ABCDEFG&", "ABCDEF&G", "ABCDE&FG", "ABCD&EFG", "&ABCDEFG", "ABCDEFGH", "ABCD&EFG", "&ABCDEFG", "ABCDEFGH"}
 
 func TestCheck(t *testing.T) {
 	st := NewShortToken(8)
-	require.NoError(t, st.Check(tokens[0]))
-	require.NoError(t, check1(tokens[0], 8))
-	require.EqualError(t, st.Check(tokens[1]), "wrong token alphabet")
-	require.EqualError(t, check1(tokens[1], 8), "incorrect symbol: '&'")
-
+	stO := newBShortTokenOrig(8)
+	stA := newBShortTokenAlt(8)
+	require.NoError(t, st.CheckAlphabet(tokens[0]))
+	require.NoError(t, stO.CheckAlphabet(tokens[0]))
+	require.NoError(t, stA.CheckAlphabet(tokens[0]))
+	require.EqualError(t, st.CheckAlphabet(tokens[1]), "illegal base64 data at input byte 7")
+	require.EqualError(t, stO.CheckAlphabet(tokens[1]), "illegal base64 data at input byte 7")
+	require.EqualError(t, stA.CheckAlphabet(tokens[1]), "illegal base64 data at input byte 7")
 }
 
 func BenchmarkTokenCheck(b *testing.B) {
 	st := NewShortToken(8)
-	for range b.N {
+	for b.Loop() {
 		for _, t := range tokens {
-			st.Check(t)
+			st.CheckAlphabet(t)
 		}
 	}
 }
 
-func BenchmarkTokenCheck1(b *testing.B) {
-	_ = NewShortToken(8)
-	for range b.N {
+func BenchmarkTokenCheckOrig(b *testing.B) {
+	st := newBShortTokenOrig(8)
+	for b.Loop() {
 		for _, t := range tokens {
-			check1(t, 8)
+			st.CheckAlphabet(t)
 		}
 	}
 }
 
-func check1(token string, l int) error {
-	// check length
-	if len(token) != l {
-		return errors.New("wrong token length")
+func BenchmarkTokenCheckAlt(b *testing.B) {
+	st := newBShortTokenAlt(8)
+	for b.Loop() {
+		for _, t := range tokens {
+			st.CheckAlphabet(t)
+		}
 	}
+}
 
-	for _, s := range token {
-		if !((s >= 'A' && s <= 'Z') || (s >= 'a' && s <= 'z') || (s <= '0' && s >= '9') || s == '_' || s == '-') {
-			return fmt.Errorf("incorrect symbol: '%c'", s)
+type testSToken struct{ length int }
+
+func newBShortTokenOrig(length int) ShortToken { return &testSToken{length: length} }
+
+func (s *testSToken) Get() string { return "AAAAAAAA" }
+
+func (s *testSToken) CheckLength(_ string) error { return nil }
+
+func (s *testSToken) CheckAlphabet(sToken string) error {
+	if _, err := base64.URLEncoding.DecodeString(sToken + "AAAA"[:4-s.length%4]); err != nil {
+		return err
+	}
+	return nil
+}
+
+type testSTokenA struct{}
+
+func newBShortTokenAlt(_ int) ShortToken { return &testSTokenA{} }
+
+func (s *testSTokenA) Get() string { return "AAAAAAAA" }
+
+func (s *testSTokenA) CheckLength(sToken string) error { return nil }
+
+var urlEncodingBytes = []byte(urlEncodingString)
+
+func (s *testSTokenA) CheckAlphabet(sToken string) error {
+	for i, s := range sToken {
+		if !bytes.ContainsRune(urlEncodingBytes, s) {
+			return base64.CorruptInputError(i)
 		}
 	}
 	return nil
